+++ src/main/java/codesquad/service/UserService.java
@@ -12,35 +12,34 @@ import java.util.List;
 
 @Service("userService")
 public class UserService {
-	@Resource(name = "userRepository")
-	private UserRepository userRepository;
+    @Resource(name = "userRepository")
+    private UserRepository userRepository;
 
-	public User add(User user) {
-		return userRepository.save(user);
-	}
+    public User add(User user) {
+        return userRepository.save(user);
+    }
 
-	@Transactional
-	public User update(User loginUser, long id, User updatedUser) {
-		User original = findById(loginUser, id);
-		original.update(loginUser, updatedUser);
-		return original;
-	}
+    @Transactional
+    public User update(User loginUser, long id, User updatedUser) {
+        User original = findById(loginUser, id);
+        original.update(loginUser, updatedUser);
+        return original;
+    }
 
-	public User findById(User loginUser, long id) {
-		return userRepository.findById(id)
-				.filter(user -> user.equals(loginUser))
-				.orElseThrow(UnAuthorizedException::new);
-	}
+    public User findById(User loginUser, long id) {
+        return userRepository.findById(id)
+                .filter(user -> user.equals(loginUser))
+                .orElseThrow(UnAuthorizedException::new);
+    }
 
-	public List<User> findAll() {
-		return userRepository.findAll();
-	}
+    public List<User> findAll() {
+        return userRepository.findAll();
+    }
 
-	public User login(String userId, String password) throws UnAuthenticationException {
-		User authorizedUser = userRepository.findByUserId(userId).orElseThrow(UnAuthenticationException::new);
-		if (!authorizedUser.matchPassword(password)) {
-			throw new UnAuthenticationException();
-		}
-		return authorizedUser;
-	}
+    public User login(String userId, String password) throws UnAuthenticationException {
+        return userRepository
+                .findByUserId(userId)
+                .filter(user -> user.matchPassword(password))
+                .orElseThrow(UnAuthenticationException::new);
+    }
 }
