+++ src/test/java/codesquad/web/LoginAccpetanceTest.java
@@ -8,10 +8,12 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.client.TestRestTemplate;
+import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.test.context.junit4.SpringRunner;
-import support.utils.HtmlFormDataBuilder;
+import org.springframework.util.MultiValueMap;
+import support.test.HtmlFormDataBuilder;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -19,47 +21,47 @@ import static org.assertj.core.api.Assertions.assertThat;
 @RunWith(SpringRunner.class)
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 public class LoginAccpetanceTest {
-	private static final Logger log = LoggerFactory.getLogger(UserAcceptanceTest.class);
+    private static final Logger log = LoggerFactory.getLogger(UserAcceptanceTest.class);
 
-	@Autowired
-	private TestRestTemplate template;
+    @Autowired
+    private TestRestTemplate template;
 
-	@Autowired
-	private UserRepository userRepository;
+    @Autowired
+    private UserRepository userRepository;
 
-	@Test
-	public void loginForm() throws Exception {
-		ResponseEntity<String> response = template.getForEntity("/users/login", String.class);
-		assertThat(response.getStatusCode()).isEqualTo(HttpStatus.OK);
-		log.debug("body : {}", response.getBody());
-	}
+    @Test
+    public void loginForm() throws Exception {
+        ResponseEntity<String> response = template.getForEntity("/users/login", String.class);
+        assertThat(response.getStatusCode()).isEqualTo(HttpStatus.OK);
+        log.debug("body : {}", response.getBody());
+    }
 
-	@Test
-	public void login_sucess() throws Exception {
-		HtmlFormDataBuilder builder = HtmlFormDataBuilder.urlEncodedForm();
-		String userId = "javajigi";
-		String password = "test";
-		builder.addParameter("userId", userId);
-		builder.addParameter("password", password);
+    @Test
+    public void login_sucess() throws Exception {
+        String userId = "javajigi";
+        HttpEntity<MultiValueMap<String, Object>> request = HtmlFormDataBuilder.urlEncodedForm()
+                                                                                .addParameter("userId", userId)
+                                                                                .addParameter("password", "test")
+                                                                                .build();
 
-		ResponseEntity<String> response = template.postForEntity("/users/login", builder.build(), String.class);
+        ResponseEntity<String> response = template.postForEntity("/users/login", request, String.class);
 
-		assertThat(response.getStatusCode()).isEqualTo(HttpStatus.FOUND);
-		assertThat(userRepository.findByUserId(userId).isPresent()).isTrue();
-		assertThat(response.getHeaders().getLocation().getPath()).isEqualTo("/users");
-	}
+        assertThat(response.getStatusCode()).isEqualTo(HttpStatus.FOUND);
+        assertThat(userRepository.findByUserId(userId).isPresent()).isTrue();
+        assertThat(response.getHeaders().getLocation().getPath()).isEqualTo("/users");
+    }
 
-	@Test
-	public void login_fail() throws Exception {
-		HtmlFormDataBuilder builder = HtmlFormDataBuilder.urlEncodedForm();
-		String userId = "javajigi";
-		String password = "test";
-		builder.addParameter("userId", userId);
-		builder.addParameter("password", password);
+    @Test
+    public void login_fail() throws Exception {
+        String userId = "javajigi";
+        HttpEntity<MultiValueMap<String, Object>> request = HtmlFormDataBuilder.urlEncodedForm()
+                                                                                .addParameter("userId", userId)
+                                                                                .addParameter("password", "test")
+                                                                                .build();
 
-		ResponseEntity<String> response = template.postForEntity("/users/login", builder.build(), String.class);
+        ResponseEntity<String> response = template.postForEntity("/users/login", request, String.class);
 
-		assertThat(response.getStatusCode()).isEqualTo(HttpStatus.OK);
-		assertThat(userRepository.findByUserId(userId).isPresent()).isFalse();
-	}
+        assertThat(response.getStatusCode()).isEqualTo(HttpStatus.OK);
+        assertThat(userRepository.findByUserId(userId).isPresent()).isFalse();
+    }
 }
