+++ src/test/java/support/test/HtmlFormDataBuilder.java
@@ -1,4 +1,4 @@
-package support.utils;
+package support.test;
 
 import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpHeaders;
